+++ threedi_modelchecker/simulation_templates/__init__.py
@@ -1,3 +1,3 @@
+from .exceptions import *  # NOQA
 from .extractor import *  # NOQA
 from .models import *  # NOQA
-from .exceptions import *  # NOQA