+++ threedi_modelchecker/simulation_templates/utils.py
@@ -1,4 +1,6 @@
-from typing import Any, List, Dict
+from typing import Any
+from typing import Dict
+from typing import List
 
 
 def parse_timeseries(